Moved event listeners into update.js. Changed -send buttons to -btn to have more flexibility.
