+++ src/js/utils/utils.js
@@ -20,3 +20,39 @@ export const alphaSortWithKey = (key) => {
         return 0;
     }
 }
+
+/**
+ * Append the options to a selector
+ *
+ * @param  {object} selector    The selector to append options to
+ * @param  {string} values      The array of value objects to append
+ * @param  {string} textKey     The key for the text of the option
+ * @param  {string} valueKey    The key for the value option
+ * @return {void}
+ */
+export const appendOptionsToSelect = (selector, values, textKey, valueKey) => {
+    clearSelector(selector);
+    values
+        .sort(alphaSortWithKey(textKey))
+        .forEach((val) =>  {
+            const option = document.createElement('option');
+            option.text = val[textKey];
+            option.value = val[valueKey];
+            selector.appendChild(option);
+        });
+}
+
+/**
+ * Clear the selector of all options but the first.
+ *
+ * @param  {object} selector    The selector you want to clear
+ * @return {void}
+ */
+export const clearSelector = (selector) => {
+    const options = selector.querySelectorAll('option');
+    options.forEach((option, i)    =>  {
+        if (i > 0) {
+            option.remove();
+        }
+    });
+}
