+++ src/templates/layout/form-multiple.pug
@@ -11,7 +11,7 @@ form.d-flex.flex-row
             a.hidden.btn-close-form(id="btn-close-form-"+id, data-input-name=id)
                 include ../icons/chevron-up
             .button-save.mt-3.hidden
-                button.bg-transparent.border-0.p-0.text-primary(id=id + '-send', type="submit")
+                button.bg-transparent.border-0.p-0.text-primary(id=id + '-btn', type="submit")
                     include ../icons/save
                 .loader(id=`${id}-loader`)
