+++ src/templates/mixins/button-save-download.pug
@@ -1,4 +1,4 @@
 mixin button-save-download(id)
     .position-absolute.end--5.top-0
-        button.bg-transparent.border-0.p-0.text-primary(id=id+'-send', type="submit")
+        button.bg-transparent.border-0.p-0.text-primary(id=id+'-btn', type="submit")
             include ../icons/donwload-solid
