+++ src/templates/mixins/button-save.pug
@@ -1,5 +1,5 @@
 mixin button-save(id)
     .button-save.position-absolute.end--5.top-0.hidden
-        button.bg-transparent.border-0.p-0.text-primary(id=id+'-send', type="submit")
+        button.bg-transparent.border-0.p-0.text-primary(id=id+'-btn', type="submit")
             include ../icons/save
         .loader(id=id+'-loader')
