+++ src/templates/parts/admin/forms/wipe.pug
@@ -15,5 +15,5 @@ block form
     div.position-relative
         input.w-100.input-text#wipe-input(placeholder='Password', name='wipe')
         .button-save.position-absolute.end--5.top-0.hidden
-            button.bg-transparent.border-0.p-0.text-primary(id='wipe-send', type="submit")
+            button.bg-transparent.border-0.p-0.text-primary(id='wipe-btn', type="submit")
                 include ../../../icons/play
